requirements.txt
Flask
Flask-Login
easyocr
opencv-python-headless
requests
python-dotenv
Pillow
PyYAML 